refactor: Move logging stub creation to tests/helpers subdir
Will be needed by more than one command test
